requirements.txt
python-telegram-bot
requests
pandas
networkx
matplotlib
